fix root font sizing and font loading
Problem 1:

The custom fonts are loaded in the _app.tsx and the css
variables are defined on the layout div. This means that if these
won't be available in the html, body scope and the font wont be applied.

Solution: Move the font loading into the .layout div. It will be
inherited by all children.

Problem 2:

The responsive font sizes need to be defined on html, and body so the
rem is set properly. (rem won't be set if the size is defined later)

Solution: Keep the font sizes on html, body and move the font definition
to the .layout div.

Problem 3: It's tempting to move all of this into _document.tsx, but
that is only rendered on the server — no window object — and react
code won't be run outside of the <Main /> component.
https://nextjs.org/docs/pages/building-your-application/routing/custom-document